+++ lib/l10n/app_localizations_es.dart
@@ -316,6 +316,9 @@ class L10nEs extends L10n {
   String get detMeetingLinkCopied => 'Enlace de la reunión copiado';
 
   @override
+  String get detAttachments => 'Adjuntos';
+
+  @override
   String get detOpenInCloud => 'Abrir en la nube';
 
   @override